# ScFrameWork/docs/openapi-starter.json
{
  "components": {
    "schemas": {
...
      "StoredBlob": {
        "properties": {
          "key": {
            "type": "string"
          },
          "size": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "session": {
        "description": "같은 브라우저의 로그인 세션. HttpOnly 쿠키를 Authorize 입력으로 설정하지 않는다.",
        "in": "cookie",
        "name": "JSESSIONID",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "sc-starter-app",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/storage-demo": {
      "post": {
        "operationId": "upload",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": ["file"],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StoredBlob"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["storage-demo-controller"]
      }
    },
    "/api/storage-demo/{key}": {
      "delete": {
        "operationId": "delete",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": ["storage-demo-controller"]
      },
      "get": {
        "operationId": "download",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["storage-demo-controller"]
      }
    }
  },
  "security": [
    {
      "session": []
    }
  ],
  "servers": [
    {
      "description": "같은 출처의 실행 서버",
      "url": "/"
    }
  ]
}